Add rules for building guest VM combined ramdisk
The guest VM ramdisk is a combination of "base ramdisk" (containing
the init process and test binaries/resources) and a "vendor ramdisk"
(containing kernel modules).

This CL adds build rules that concatenate the base and vendor ramdisks.
This can be done as a simple `cat x1 x2` because both LZ4 and CPIO
formats are concatenable.

One ramdisk per target architecture is generated because genrules cannot
be specialized using the 'arch' field.

Test: m virt_hostside_tests_initramfs-{arm64,x86_64}
Change-Id: I7f52226ea68b8bfb497e481538eb35267578c38f