Refactor VirtualizationHostTestCases
The init ramfs used by the test is now built using the soong module type
`android_filesystem`. This not only simplifies the build script, but
also verifies that the new build system construct correctly creates the
ramdisk image.

In addition, only the test data (kernel, initramfs, and the server) that
match with the target architecture are copied.

Lastly, the test now uses crosvm from inside the virt APEX.

Bug: 178978059
Test: atest VirtualizationHostTestCases
Change-Id: Ibacc66fc5842f950c9ea16001e62bd44ed9b867c