linker.config.pb for microdroid is built using the linker_config module
Note that we don't use system_linker_config module which is for the host
side (Android). We can't use the same module because we have different
lists of provideLibs across Android and microdroid. Note that only the
provideLibs are automatically generated. requiredLibs is still
hard-coded and is copied from system/core/rootdir/etc and edited to
remove the files that will never be on microdroid.

Bug: 181093750
Test: m microdroid
Change-Id: I500e9e5ce73773751951ad77a1d4f3c663ea2510