microdroid: each partition in payload has size in it
When creating the composite disk image for payload, use FILL_WITH_SIZE
option, which adds padding to fill the holes in disk image and the size
of original backing file as well.

In guest OS, readers can look up the last 4 bytes to get the actual size
of host files.

Bug: 187783952
Test: MicrodroidHostTestCases
Change-Id: I072873779df4737448112e262fc1b051c0229ea8