microdroid_manager: waits for a config file
Ideally, microdroid_manager should start zipfuse only if necessary and
access the VM config file from the APK via fusefs when it's ready.

But for now zipfuse doesn't report when it's ready to access files in a
zip file. So, microdroid_manager should wait a while to access the
config file when it is from the APK.

Bug: 189301496
Test: MicrodroidHostTestCases
Change-Id: I0dc304a8f135f52a846fc1bb6f4e476f6162697a