Specialize the KM context to use encrypted key blobs
These key blobs are intended to be exported outside the VM, without need
for further encryption. This will allows a limited form of persistence
for VMs between boots of the same code.

The authorization set is left in plain text when using the existing
encrypted key blob format. If that causes a problem then we'll have to
create a custom format that also protects those details.

Bug: 190578423
Test: atest MicrodroidHostTestCases
Change-Id: I47a0f80e2137e189634b77c0b4aafb32d002be50