Merge AuthFsHostTest into MicrodroidTestCase
In the meantime, in order to reduce the amount of testing time, boot the
VM once in AuthFsTestCase. This is possible through
@BeforeClassWithInfo.

The new DeviceRunner class intends to generalize shell command execution
to ITestDevice. Later when the VM can be represented as an ITestDevice,
we can easily run a command in the same way.

Bug: 191056545
Bug: 182478337
Test: atest MicrodroidHostTestCases

Change-Id: I680a9ef00c345a2f4d7773412c7f0ebd767aab15