Integration test for CompOS
The only test case initially is to verify that odrefresh can succeed
with dex2oat running in the VM.  The current goal is to keep the test
running.

In the longer term, when we're ready to increase the end-to-end test
coversage, we might want to figure out a better place to host the test,
e.g. possibly odsign_e2e_tests.

Bug: 192690283
Test: atest ComposHostTestCases
Change-Id: Ic1395437d8bc13c72b21e6460242ccb445be2b87