Add support for create image & start VM.
Add a make-image command to create an empty VM image.

Add --start <image> option to start a CompOS VM & then connect to it,
rather than some existing VM. Also add --log to capture the VM output
to file.

This is mostly for test purposes, and to allow me to experiment with
using IVirtualizationService from C++ before I start adding it to
odsign.

Bug: 193603140
Test: compos_key_cmd make-image <if>
Test: compos_key_cmd --log <log> --start <if> generate <blob> <key>
Test: compos_key_cmd --log <log> --start <if> verify <blob> <key>
Change-Id: I32419837ac4cf0c69180382fa7eaee591056f85d