compos: retrieve fsverity digests of outputs from authfs
In order for compsvc to know which files to sign, it needs to parse the
compliler flags to identify the FD numbers. Once the complier exits
normally, we retrieve the fs-verity digests of the output files, namely
the oat, vdex and image.

The digests are currently only printed as debug log, but with more
changes to the signing service, we'll be able to sign them easily.

Bug: 161471326
Test: ComposHostTestCases
Test: Visually checks the hex of the fs-verity digests. Looks consistent
      to the digests of the actual files.
Change-Id: I9f7b2cfbb12999b87676b6b8e1bfd199d2a75a6b