Store public key of APK to instance disk
Once a VM is started with an APK, we shouldn't accept any update of the
APK if its signer has changed.

For now, this CL can be considered a no-op because we don't accept any
update by comparing the root hash. However, in the future, when rollback
protection is supported, this change will make it possible to accept APK
updates if the public key remains the same.

Bug: 199143508
Test: atest MicrodroidHostTestCases

Change-Id: I2448faf4a8b9637571ebcc4bc49d3619129496d5