Fix lazy service persistence for VS
As long as there is a VmInstance in existence, i.e. a client is
holding an IVirtualMachine reference, keep the virtualization service
running. Previously it would be killed by service manager when its
lazy IVirtualisationService interface was released, but that is
unhelpful to clients.

This reverts the changes in commit
7e54e2904f414369054114a0d5b9ed0e6d3a23cd that special-cased
debug_hold_vm in favor of the more general solution.

It also removes the workaround in compos_client to keep the service
from being killed.

Test: atest ComposHostTestCases
Bug: 200924402
Change-Id: I8ba465b5970746b64205e08745d3ec4c931abae8