Test CompOS more directly.
Rather than using "vm --daemonize ..." to test, which is a slightly
unrealistic environment and doesn't exercise much of our client-side
code, use composd where we can.

A side-effect of this is that we no longer need to disable selinux for
the tests.

This is a partial revert of commit
6aff4f827839325647a43d3a884e6e3d97075d54, and thus a partial re-submit
of commit a00d792a8f7f4bca3d4deae34abbbc0dcca02f2b, changing only
ComposTestCase. I'll follow up with the remanining changes to
ComposKeyTestCase, fixed, since they are higher risk and less urgent.

Bug: 200924405
Test: atest ComposHostTestCases
Change-Id: Ia56ce4b7305d14d6b759ede0d612eb3095559576