Extract and rename InstanceFiles.
It seems to be complex enough to merit its own file and not have
everything effectively public.

I renamed it to InstanceStarter, since what it does is start an
instance (possibly creating/re-creating it first). And did a little
related renaming.

This is intended as a pure refactoring; no functionality changes.

Bug: 186126194
Test: adb shell apex/com.android.compos/bin/composd_cmd
Change-Id: Ib42e1c23594d436b845c28bcfe38803e119d1106