compsvc: New compile method without cmdline arguments
The new method will supercede the old `compile_cmd` method once
everything is ready. This is the service side implement that corresponds
to the previous client API.

Note that the new API is not handling the output signature yet, because
compsvc cannot simply associate output FD numbers with file type.

Bug: 193668901
Test: atest ComposHostTestCases (with other changes)

Change-Id: I91209b9ec14418711719ad382814e153244351da