Temporarily disable testDebugMode
When testDebugMode and testMicrodroidBoots run in succession, it looks
like the run of a test is affecting the next test. Specifically, the
killing of a VM somehow cuts the adb connection to the host. Until the
root cause is fixed, let's disable the testDebugMode test. This is not a
big loss because we currently don't rely on the debug mode.

Bug: 195765441
Test: watch TH on both cuttlefish and real devices.
Change-Id: I086afd10ac0ea4c70a2908af09f8decda03e6dad