Pass log_fd directly as the backend for serial devices
Previously, the backend for serial devices was configured to be stdout
of the crosvm process and the stdout was redirected to log_fd. This was
enevitible because the --serial flag in crosvm didn't accept arbitrary
file descriptor as the backend. This also prevented us from having other
serial devices for logcat logs.

Crosvm now supports serial devices backed by file descriptors using the
/proc/self/fd/N syntax, like it does for disks. Let's make use of the
new syntax in preparation for adding more serial devices.

Bug: 200914564
Test: atest MicrodroidHostTestCases
Change-Id: I3fc160abe8cc102f189ebe6d490afd94d3ef04b0