Use a test helper executable to open files
In the test, we used a trick to open files from the shell, the pass the
FDs to fd_server. It will not work for directories.

This change replaces the trick with a test helper executable, which
supports opening directories as FD, in order to make the future test
cases that involved directory FD possible.

Bug: 203251769
Test: AuthFsHostTest
Change-Id: I77fdfa463f168bce99d1fccc91d0739bbc6ba00a