Move CompOS to /system_ext
We shouldn't be on /system since we're not generic (yet).

But we are closely coupled to /system (integrated with ART, will be
contributing to the system server classpath.

Prune the artifact path files to only the ones that are still written
to /system while we figure out what to do with them.

Test: Boots, APEXes present
Bug: 186126194
Change-Id: Id84fb7bf0fb4e0fb56e0cb896394c31e5c7b5124