Wait for logd before starting the seriallogging service
This CL fixes a bug that logcat logs from microdroid is shown very
lately (e.g. after 5 sec). This was because when the seriallogging
service which is responsible for the logcat functionality is started,
logd is not ready yet, causing logcat to crash. The service was then restarted
by init after 5 seconds.

To fix this, this CL introduces a new system property ro.logd.ready to
synchronize logd and its clients including the seriallogging service.

Bug: 206826522
Test: run microdroid and see that logcat logs are shown immediately
Change-Id: Ica7446824442c8b79546ef537044f8b2d5b5516a