Don't install files for the virt APEX onto the system partition
Previously, since the virt APEX is configured as non-updatable, the
build system enabled the symlink optimization where files are installed
directly to the system partition and the APEX has symlinks to the files.
However, this forced us to add those files to
PRODUCT_ARTIFACT_PATH_REQUIREMENT_ALLOWED_LIST as we install the APEX
conditionally - because the inclusion of the APEX is done by a product,
not by the core build system.

This change works around the problem by marking the APEX
non-installable. Then the build system doesn't enable the symlink
optimization.

Bug: N/A
Test: watch TH

Change-Id: I020ae84defe5ce03f0fc2226f2fbd42a5823861e