Force disable symlink optimization for the virt APEX
Previously, since the virt APEX is configured as non-updatable, the
build system enabled the symlink optimization where files are installed
directly to the system partition and the APEX has symlinks to the files.
However, this forced us to add those files to
PRODUCT_ARTIFACT_PATH_REQUIREMENT_ALLOWED_LIST as we install the APEX
conditionally - because the inclusion of the APEX is done by a product,
not by the core build system.

This change works around the problem by adding a new switch
`force_disable_size_optimization` and turn it on for the virt APEX.

Bug: 207336449
Test: watch TH
Test: run asit/ota/signing
Change-Id: I841b46895df036f278a85dd318f3317a351992c3