Run derive_classpath in VM
This allows us to get an accurate list of APEXes that contribute to
the classpaths, including staged APEXes. (Note that derive_classpath
itself is in an APEX, and we will run it from a staged APEX if
appropriate.)

Add com.android.sdkext (which contains derive_classpath) as an
explicit dependency; also fixed formatting while I was at it.

Remove the old initializeClasspaths method as it is now redudant.

Tweaked build.prop values to allow derive_classpath to run; more work
is needed though (b/189164487).

Bug: 200808307
Test: composd_cmd test-compile
Change-Id: Iae677274ebac1cdf721c5497c1d82eb22d2277db