Use --force-compile for tests
The test uses --force-compile outside the VM, so we also need to
inside the VM. But we want --compile for the normal case of
compilation, so we only produce exactly the required artifacts.

That means we have to pass whether we are testing or not from the host
to guest, so lots of plumbing is needed.

(Also figured out how to solve a problem with type deduction that the
me of a few months ago was baffled by.)

Bug: 211458160
Test: atest ComposTestCase
Change-Id: Ie7e36bc3fab40744e9f9ea5e5928126899d9060d