Test vm secrets are different between instances
Different instances of VMs must be provisioned with different secrets,
even when they are otherwise identical. This should from the very start
of the VM and continue to be true after things have been loaded in the
VM. This test checks that the payloads in two identical microdroid VMs
instances are provisioned with different secrets.

There is a slight chance for the test to fail by chance, but it is
cryptographically insignificant by design.

Bug: 214231981
Test: atest MicrodroidTestApp
Change-Id: I488c7370ad901a73a520bc3fa548d4b9b7959d19