Set a property to reflect the app debuggability
When loading the payload app, microdroid_manager checks the
debuggability and uses that as a DICE input. Make this value available
to apps that also make decisions based on debugability, such as compos, don't
have to try and replicate the logic.

This will make it easier to tweak the definition if debuggability and
keep it consistent.

Bug: 219740340
Test: atest ComposHostTestCases
Test: atest MicrodroidTests
Change-Id: I770ae615a958dd6939925c400336a2550a3d950d