Enforce mk_payload config and output files to be in cwd
Relative paths do not work unless this is the case. This is because the
paths used when reading the files and writing to the composite disk are
the same and based on cwd. We could update image_aggregator at a later
point if needed but for now it is sufficient to let people know what
to expect.

Bug: b/221887571
Test: Run manually and see failure.
Change-Id: I598e332a8b22503d6f6a2ff67517225b86c43084