Make exporting tombstone (out of VM) configurable
VM owner can set export_tombstones to false in assets/vm_config in
payload apk to disable exporting the tombstones out of VM. This is also
dm-verity protected so only the owner of the VM can make the change. By
default it is enabled.

Test: atest MicrodroidHostTestCases (particularly
testTombstonesAreBeingForwarded passes when export_tombstones is true).
Manually tested that it fails when export_tombstones = false
Bug: 227443903

Change-Id: I82325a9ffc57d4d4825d51d9d41ebb767e6c1f6f