apkdmverity: Using LOOP_CONFIGURE
apkdmverity can attach a file to a loop device using LOOP_CONFIGURE
instead of LOOP_SET_FD/LOOP_SET_STATUS64 because it's more efficient.
The reason we used the latter was that LOOP_CONFIGURE didn't work but
the actual reason is enabling Direct-IO for not-aligned small file (in
this case, IDSIG file).

Now, we explicitly do not turn on Direct IO when attaching IDSIG file
to a loop device.

This is a partial revert of 7ce2e5352866e44e9499cf4dfbdee8c72ab8a7a4.

Bug: 191344832
Test: atest MicrodroidTestApp
Test: atest apkdmverity.test
Change-Id: I53a9da6d1d981b0d64a75bd362b09063570e7a68