Fe-enable test with fixes
Previously if enableAdbRoot() failed we would leave the VM running
(because sMicrodroid was still null), leading to a cascade of failures
in subsequent tests.

This doesn't resolve the issue that enableAdbRoot() sometimes seems to
fail on Cuttlefish (probably due to nested virtualization slowdown),
but should mean it is a normal flake rather than a blocker.

Bug: 229823049
Test: atest AuthFsHostTest
Test: Fake adb root failure, observe VM torn down
Change-Id: I3a365a557faa99babdf404b430d5f412f5895378