Enable log for debug_mode == NONE
Logs are now captured from VMs with debug_mode == NONE. This is to ease
debug problems on user builds (i.e. production devices). This is
considered safe as the isolated compilation is the only client of AVF
and it doesn't reveal its secret via log, which actually nobody should
do.

Bug: 216538816
Test: run isolated compilation on user build of P22.
Change-Id: I9f2e3d9f6762f3a82141c9ebbc7f1e2c39d0a4ed