Detect Microdroid hangup during boot
If Microdroid, for some reason, fails to start the payload within 5
secs (or 30 secs on a nested virt environment), such an event is
detected as a hangup and `onError` is invoked with an appropriate error
code and message. The event however doesn't kill the VM immediately,
because the VM might just be experiencing a temporary slowdown, perhaps
due to the high workload on the host side. So it's up to the client to
handle the event.

If the event is not handled by the client, the system again waits the
same amount of time, and if the payload is still not started, it
triggers the event again, and reset the timer, up to 5 times, in which
case the system automatically kills the VM without the explicit request
from client. Finally, `onDied` is called.

Bug: 222228861
Test: I couldn't make a test for this because it was impossible to
intentionally make the hang by a test. Instead, I confirm that `onError`
is called and the VM eventually is killed when I edited the timeout
value to a very small number (e.g. 100ms).

Change-Id: I53f232d0b609e6e8a429d996c7d6fdd0b37e7b4c