WIP: Remote attestation keys for VMs
A method is added to the IVirtualizationMachine service to allow VMs to
request a remotely attested certificate chain for a public key. That
service is implemented by calling into "rkpvm" which will service the
request.

rkpvm is currently implemented as a normal process but it will need to
migrate into a VM. Building a bare metal VM on the rust base looks to be
the best approach.

The request is routed through VirtualizationService in the untrusted
host but only public information needs to be exchanged so this doesn't
pose a problem. Only the client authenticates to rkpvm, with its DICE
chain, as rkpvm needs to know who it is certifying the keys for, but the
client doesn't need to know it's actually talking rkpvm as it is only
denial-of-service if a malfunctioning remote attestation key is
returned.

A fake attestation certificate chain is generated, consisting of a
single certificate signed by a random key. This will be replaced with an
RKP-provisioned key.

Test: WIP
Change-Id: I4d9043de1f6eaffcbfd157234f87bd02f46e190f