WIP: Remote attestation keys for VMs
A method is added to the IVirtualizationMachine service to allow VMs to
request a remotely attested certificate chain for a public key. That
service is implemented by calling into "rkpvm" which will service the
request.

rkpvm is currently implemented as a normal process and expected that Rialto
would help it migrate into a VM. But that plan will need to be
replaced..

The request is routed through VirtualizationService in the untrusted
host but only public information needs to be exchanged so this doesn't
pose a problem. Only the client authenticates to rkpvm, with its DICE
chain, as rkpvm needs to know who it is certifying the keys for, but the
client doesn't need to know it's actually talking rkpvm as it is only
denial-of-service if a malfunctioning remote attestation key is
returned.

A fake attestation certificate chain is generated, consisting of a
single certificate signed by a random key. This will be replaced with an
RKP-provisioned key.

The rkpclient test VM uses the new interface to get a key signed. It
requires an apex_available update for the diced HAL because somehow
udroid bypasses those checks. I've been using the script below to run
the client test installing the updated APEX.

```
set -euxo pipefail

TEST_ROOT=/data/local/tmp/rkpvm

adb wait-for-device
sleep 1
adb shell mkdir -p $TEST_ROOT
adb shell rm -f $TEST_ROOT/instance.debug.img
adb shell /apex/com.android.virt/bin/vm run-app
/apex/com.android.virt/app/RkpClientPayloadApp@AOSP.MASTER/RkpClientPayloadApp.apk
$TEST_ROOT/MicrodroidTestApp.apk.idsig $TEST_ROOT/instance.debug.img
assets/vm_config.json --debug full --protected
```

Test: WIP
Change-Id: I4d9043de1f6eaffcbfd157234f87bd02f46e190f