WIP: Remote attestation keys for VMs
A method is added to the IVirtualizationMachine service to allow VMs to
request a remotely attested certificate chain for a public key. That
service is implemented by calling into "rkpvm" which will service the
request.

rkpvm is currently implemented as a normal process and expected that Rialto
would help it migrate into a VM. But that plan will need to be
replaced..

The request is routed through VirtualizationService in the untrusted
host but only public information needs to be exchanged so this doesn't
pose a problem. Only the client authenticates to rkpvm, with its DICE
chain, as rkpvm needs to know who it is certifying the keys for, but the
client doesn't need to know it's actually talking rkpvm as it is only
denial-of-service if a malfunctioning remote attestation key is
returned.

The attestation key is hackily generated with the help of keystore to
get a Google-signed root. The key could be used to attest that, assuming
the device is not compromised, the VM is running on an Android device...

The rkpclient test VM uses the new interface and exchanges data over the
negotiated channel.

Oh, this won't work alone, there's also a bunch of hacks all over the
tree, particularly for apex_available.

Test: WIP
Change-Id: I4d9043de1f6eaffcbfd157234f87bd02f46e190f