Add kexec and crashdump
A simple implementation of kexec is added to Microdroid. It will be
executed during the boot to load the crashdump kernel in a reserved
memory. When a kernel panic occurs, the kernel will boot to the
crashdump kernel and execute the crashdump program as the init. The
crashdump copies /proc/vmcore (which contains memorydump of just before
the soft reboot) to the target device backed by a file in the host
(Android).

The kexec implementation is (intentionally) not complete. It doesn't
implement all the features that the upstream kexec_tools supports. It
implements only the feature required for loading the crashdump kernel
and initrd.

Bug: 237381237
Bug: 237358006
Bug: 237380224
Bug: 237381762
Bug: 237380607
Test: this isn't automated yet.
Build AVF using a custom kernel where these configs are enabled.
CONFIG_KEXEC_FILE=y
CONFIG_CRASH_DUMP=y
CONFIG_PROC_VMCORE=y
CONFIG_DEVTMPFS=y

Then start a Microdroid VM, and adb into it. Execute the following
command to load the crashdump kernel and force trigger a panic.

/system/bin/kexec \
  /system/etc/microdroid_crashdump_kernel \
  /system/etc/microdroid_crashdump_initrd.img \
  "rdinit=/bin/crashdump 8250.nr_uarts=4 printk.devkmsg=on earlycon=uart8250,mmio,0x3f8"

echo c > /proc/sysrq-trigger

Change-Id: If964bb2a7b2e444cd57870b281405cd522713b20