Add crashdump kernel
microdroid_crashdump_kernel is a very small kernel that is executed when
a kernel panic occurs. Since it's only use case is to run the crashdump
binary, the kernel is very minimal. In addition, it has to be as small
as possible because the kernel has to be loaded (using kexec) in the VM
RAM before a panic occurs.

To support these needs, it is built with most configuration options
turned off, except for the followings needed for dumping the RAM:

* CONFIG_CRASH_DUMP
* CONFIG_PROC_VMCORE
* CONFIG_RELOCATABLE

This is basically following the guideline at [1].

In addition, the following device drivers are enabled.

* CONFIG_SERIAL_8250 // for use as the earlycon
* CONFIG_VIRTIO_CONSOLE // for use as the dump target

[1] https://www.kernel.org/doc/html/latest/admin-guide/kdump/kdump.html#dump-capture-kernel-config-options-arch-independent

These kernel prebuilts are for now locally built. It will soon be
replaced with the prebuilts from the build bot when the kernel config is
uploaded to the Android common kernel repo.

Bug: 237380607
Test: N/A (follow-up CLs needed)

Change-Id: I238e1dbb74c0fc47fd4bb926b7d09caef4b90e78