vmbase: Replace libc with musl's memset/memcpy
Rust depends on libc for implementation of memory intrinsics, memset and
memcpy. So far we have met these dependencies by linking in the whole
bionic library. This is wasteful as we do not actually use any other
libc functions in vmbase.

There are several potential implementations that we could use instead:
 1) only memcpy/memset from bionic - optimized for arm64 but includes
    extra runtime checks (fortify), must implement its asm callbacks;
    can be risky if the checks interfere with the baremetal environemnt,
 2) memcpy/memset from musl - optimized for arm64, no dependencies;
 3) compiler-rt crate - designed for Rust but not optimized for arm64,
    potentially a better long-term solution.

Since memcpy/memset performance will be important for baremetal kernels,
let's start with 2), optimized implementations from musl.

This saves 34KiB in vmbase and all kernels derived from it.

Test: m vmbase_example
Change-Id: I4442f0f3b218bc6d6d8663e7d12695dba7da93cd