Shut down CompOS VM after compilation completes
Add a method to CompOsInstance to consume the instance and shut it
down cleanly, allowing logs to be written. Call this when compilation
ends, before calling back to our client (since they might then drop
their reference to us, which could tear everything down prematurely).

I also created a method to combind the calls to quit() and
wait_for_shutdown() to reduce duplication.

Bug: 236581575
Test: composd_cmd test-compile, observe logs
Change-Id: I34a227db0e377cb9aded57461379e17d0574ea0c