Build test APKs against API stubs
Modify the test APKs and their dependencies to build against our stub
library, instead of the implementation, and switch to the
system_current SDK rather than platform APIs. (We still use platform
APIs for the native payload code though.)

Modify the implementation library to statically link
HypervisorProperties, since it's inaccessible to app code, with jarjar
rules to rename it and remove the unneeded sysprops.

Fix a longstanding typo.

Bug: 242828259
Test: atest MicrodroidTests
Change-Id: I836a45e225661f0bb29f48f8753a24b4de202cc1