Report breakdown of Microdroid userspace memory usage
Dump the list of running processes in Microdroid using 'ps -A' and their
memory usage statistics from '/proc/<pid>/smaps_rollup', and report them
as metrics. Do the same for '/proc/meminfo'.

This requires root, so it is done from the host-side Microdroid tests.
Not ideal approach as the reported stats include adbd and its children,
but it is a useful set of data nonetheless.

Test: atest MicrodroidHostTestCases
Change-Id: I24368f7e647a083488ff0d9f485e618876d27f3a