Revert "Run avmdtool_tests as rust_test"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/a3d7004f-583c-4311-86cc-c59b59b0a284). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If9d2d595f8f37315aa82dd1ab714d48c657ab5ff