Revert "Run avmdtool_tests as rust_test"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/ab1a46c4-d70a-4f06-95d2-058d8981c430). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia547b14b5c13a9cacfd9e4426bcb3a88bddeac7b