Revert "Run avmdtool_tests as rust_test"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/2c202315-35ed-4ee7-bbeb-9daa8cebdd6b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie83660c44760b24a6744e1dc4e7c82c3666b6cbe