Expose VirtualMachine API as @SystemApi
This simply marks all the public members of the existing internal
platform API as @SystemApi. Update system-current.txt to match.

Clearly some modification will be required; this is mostly a proof of
concept, useful to enable testing building against the stub APIs that
are now generated. There is no intention of submitting it as-is.

Bug: 242828259
Test: atest MicrodroidTests
Change-Id: Ibfffbb221f3ac37e03d5cfdbc9208428eeadbb9f