Migrate VirtualMachine API to @SystemApi
Mark all the public members of the existing internal platform API as
@SystemApi. Then make various changes, mostly to make API lint happy:
- Replace enums with integer constants. Map the AIDL constants to the
  API constants. (Some of this was already done in
  https://r.android.com/2203781.)
- Make connectToVsockServer synchronous rather than using Future.
- Make VirtualMachine AutoCloseable (close() == stop()).
- Add VirtualMachine getters to match the builder's setters.
- Lots of nullability declarations.
- Make VirtualMachineCallback an abstract class instead of an
  interface.

Update system-current.txt to match.

Modify the demo app and the test APK and dependencies to build against
our stub library, instead of the implementation, and switch to the
system_current SDK rather than platform APIs. (We still use platform
APIs for the native payload code though.)

Bug: 243512107
Bug: 243512115
Test: atest MicrodroidTests MicrodroidHostTestCases
Test: Install & run demo app
CTS-Coverage-Bug: 244561836
Change-Id: Ibfffbb221f3ac37e03d5cfdbc9208428eeadbb9f