pvmfw: Handle MMIO guard inject exceptions
Handle the fault the hypervisor injects when we access MMIO for the
first time by sharing back the page with the host. Currently, pvmfw
receives the exception when it tries to log to the UART, resulting in
its handler being run, which then tries to write to the UART (generating
another exception) and it gets stuck in a loop.

Bug: 237371962
Test: m pvmfw && access UART
Change-Id: Ia384f2e3a15b40b5e67cfafd50e571848fd5a126