Revert "Use apexes on device APEX for testing"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/0b767524-2613-45a6-bd62-ee78b26eca38). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7ea04ee625fc6c536e236397e56401d58ca7760d