Revert "Use apexes on device APEX for testing"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/5dd7ba46-0a4b-4d53-a43e-ac985603fa72). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I965d8dd106e721450f415bc6b26a7783b8c8eec5