Revert "Use apexes on device APEX for testing"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/b6dabe31-acd7-442d-b1a4-717ea9e55791). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I46eca879bbcaa06157ad76f4b2992a360a0f3369