Revert "Fix skipping testBootEnableAndDisablePKVM (take 2)"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/30f75788-2f32-461e-93b5-a74ec981d89c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If3fe75f405fbdb42070bb1d3f3f8352b43a283f3