Revert "Fix metric reporting from AVFHostTestCase"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/30f75788-2f32-461e-93b5-a74ec981d89c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9987be07d5f2130287444a6db2bb95db0846d51f