Revert "vmbase: Make vmbase_elf_defaults android_arm64-only"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/86da8ed9-eec1-4b64-b92e-3100870db0fe). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4e5af3075b28a899f04b0c874adbcf8809409689