Revert "apkverify:Test apk digest computation"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/f1c81ef6-eded-4350-a139-581d96ce825c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I883374eb4968e9e3979ea470f4ce2ad13b1ce585