Revert "apkverify:Test apk digest computation"
Revert "apkverify:Test apk digest computation"

This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/f73be585-370d-4cc4-a1fd-a6400a14b821). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3af4298f5023d45698c43832e0648a9c83c2ecdd